% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_conver_demo.R
\name{count_conver_demo}
\alias{count_conver_demo}
\title{Title}
\usage{
count_conver_demo(
  expr,
  specie = c("human", "mouse", "rat"),
  genetype = c("hgnc_symbol", "mgi_symbol", "ensembl_gene_id"),
  from
)
}
\arguments{
\item{expr}{the expression matrix, rowname is the gene name and coulmn is the sample name}

\item{specie}{select specie from human, mouse and rat}

\item{genetype}{which kinds of gene name in expr select from hgnc_symbol, mgi_symbol and ensembl_gene_id.}

\item{from}{the count type of expr. select from count, fpkm, rpkm, tpm}
}
\description{
Title
}
