% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_demo.R
\name{deseq2_demo}
\alias{deseq2_demo}
\title{Title}
\usage{
deseq2_demo(
  exp,
  group,
  compared,
  merge = F,
  p.name = "pvalue",
  symbol = NULL,
  fc.name = "log2FoldChange",
  p.value = 0.05,
  fc.value = 0.585,
  add_expr = T,
  file.name = NULL
)
}
\arguments{
\item{exp}{expression data (if merge==T,exp need a colum to contain gene symbol).}

\item{group}{group is the vector which indicate the sample's group}

\item{compared}{compared used to identify which group vs which group (usually <conventional> - <control>). also use for multi group condition.}

\item{merge}{if expression data contain same gene symol, you can use it to merge them}

\item{p.name}{the colname of pvalue}

\item{symbol}{accordance with "merge==T", specific the colum name of symbol}

\item{fc.name}{the colname of fold change}

\item{p.value}{the threshold of pvalue}

\item{fc.value}{the threshold of fold change}

\item{file.name}{if need to write a tdata:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAbElEQVR4Xs2RQQrAMAgEfZgf7W9LAguybljJpR3wEse5JOL3ZObDb4x1loDhHbBOFU6i2Ddnw2KNiXcdAXygJlwE8OFVBHDgKrLgSInN4WMe9iXiqIVsTMjH7z/GhNTEibOxQswcYIWYOR/zAjBJfiXh3jZ6AAAAAElFTkSuQmCCxt file to save the deg result, you can assign a file name}
}
\description{
Title
}
