% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_merge_demo.R
\name{gene_merge_demo}
\alias{gene_merge_demo}
\title{Title}
\usage{
gene_merge_demo(exp, symbol, method = mean)
}
\arguments{
\item{exp}{the expression data contain one col indicate the gene name}

\item{symbol}{the colname of the gene name col}

\item{method}{how to merge the same gene count ,default as mean}
}
\description{
Title
}
