% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_demo.R
\name{go_demo}
\alias{go_demo}
\title{go analysis}
\usage{
go_demo(
  genesymbol,
  ntop = 10,
  plot = T,
  plot.name = NULL,
  file.name = NULL,
  plot.width = 7,
  plot.height = 7,
  species = c("org.Hs.eg.db", "org.Mm.eg.db"),
  ont = "ALL",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2
)
}
\arguments{
\item{genesymbol}{the enrichment gene list}

\item{ntop}{choose how many category should exhibit in barchart}

\item{plot}{if need to plot barchart}

\item{species}{choose species}
}
\value{
a list contain a datafram and a ggplot object
}
\description{
go analysis
}
