% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_demo.R
\name{scRNA_demo}
\alias{scRNA_demo}
\title{Title}
\usage{
scRNA_demo(
  path,
  specise = "human",
  min.cells = 3,
  min.features = 200,
  n.dims = 20,
  resolution = seq(0.1, 1, 0.1),
  assay = "SCT",
  interactive_mode = FALSE
)
}
\arguments{
\item{path}{the file path of single cell rna seq, every sample use the sample name as folder name}

\item{specise}{indicate the specise, human or mouse}

\item{min.cells}{minimal cell number of sample}

\item{min.features}{minimal feature number of per cell}

\item{n.dims}{the dims of pca, default as 20}

\item{resolution}{the resolution of umap, default as 0.6. if you want to specific it according to the cuttree, you can use interactive_mode to assign it at specific step.}

\item{assay}{default matrix to use. default as SCT, you can alter it to RNA.}

\item{interactive_mode}{if you want to assign cell filter parament or resolution you can alter it as TURE, or it will go through the whole pipline using default parament for all steps.}
}
\description{
Title
}
