% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea_demo.R
\name{ssgsea_deom}
\alias{ssgsea_deom}
\title{ssgsea and gsva analysis}
\usage{
ssgsea_deom(
  exp,
  pathway = NULL,
  method = "gsva",
  species = "Homo sapiens",
  category = "C2",
  subcategory = "KEGG"
)
}
\arguments{
\item{exp}{a expression data with gene symbol as rowname and sample name as colname}

\item{pathway}{the gene sets list, if pathway is NULL it will default as kegg for human}

\item{method}{method can choose from gsva or ssgsea}

\item{species}{species are name in msigdbr_species()}

\item{category}{category are name in msigdbr_collections()}

\item{subcategory}{subcategory are name in msigdbr_collections()}
}
\value{
a matrix of gsea or gsve result
}
\description{
ssgsea and gsva analysis
}
