% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_demo.R
\name{surv_demo}
\alias{surv_demo}
\title{survival analysis}
\usage{
surv_demo(expr, group)
}
\arguments{
\item{expr}{the data frame contain the expression of variates group and time, status. keep in mind, the colname of time and status must be 'time' and 'status'}

\item{group}{the colname that need to be compared group in expr}
}
\value{
survival plot
}
\description{
survival analysis
}
