% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_demo.R
\name{deseq2_demo}
\alias{deseq2_demo}
\title{Title}
\usage{
deseq2_demo(
  exp,
  group,
  compared,
  merge = T,
  p.name = "pvalue",
  fc.name = "log2FoldChange",
  p.value = 0.05,
  fc.value = 0.585,
  file.name = NULL
)
}
\arguments{
\item{exp}{expression data (if merge==T,exp need a colum to contain gene symbol).}

\item{group}{group can be multi group, and first col is sample name; second col is condition; coef, and multi used to select multi group, keep in mind, control sample priority.}

\item{compared}{compared used to identify which group vs which group (usually <conventional> - <control>). also use for multi group condition.}

\item{merge}{if expression data contain same gene symol, you can use it to merge them}

\item{p.name}{the colname of pvalue}

\item{fc.name}{the colname of fold change}

\item{p.value}{the threshold of pvalue}

\item{fc.value}{the threshold of fold change}

\item{file.name}{if need to write a txt file to save the deg result, you can assign a file name}
}
\description{
Title
}
