% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_demo.R
\name{gsea_demo}
\alias{gsea_demo}
\title{Title}
\usage{
gsea_demo(
  symbol,
  rank,
  geneset = NULL,
  go.ont = "ALL",
  GO = TRUE,
  species = c("org.Hs.eg.db", "org.Mm.eg.db"),
  KEGG = TRUE,
  n = 10,
  pathway_ID = NULL,
  pvalueCutoff = 0.05
)
}
\arguments{
\item{symbol}{gene symbol vector for the enrichment, must matched with rank}

\item{rank}{rank vector for the symbol, only matched with symbol, can be logfc or some sort index for symbol}

\item{geneset}{can specific some special gene sets}

\item{go.ont}{go.not can be "ALL", "BP", "MF" and "CC"}

\item{GO}{whether need go gene set}

\item{KEGG}{whether need kegg gene set}

\item{n}{how many pathway need to be plot}

\item{pathway_ID}{specify some interest geneset}

\item{species.go}{specify the specie}

\item{species.kegg}{specify the specie}
}
\description{
Title
}
