% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_demo.R
\name{kegg_demo}
\alias{kegg_demo}
\title{kegg analysis}
\usage{
kegg_demo(genesymbol, ntop = 10, plot = T, species = "org.Hs.eg.db")
}
\arguments{
\item{genesymbol}{the enrichment gene list}

\item{ntop}{choose how many category should exhibit in barchart}

\item{plot}{if need to plot barchart}

\item{species}{choose species}
}
\value{
a list contain a datafram and a ggplot object
}
\description{
used to execute kegg enrichment
}
