% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_demo.R
\name{limma_demo}
\alias{limma_demo}
\title{deg analysis}
\usage{
limma_demo(
  exp,
  group,
  compared,
  normalize = F,
  log2 = F,
  merge = F,
  rna.count = F,
  p.name = "P.Value",
  fc.name = "logFC",
  p.value = 0.05,
  fc.value = 0.585,
  file.name = NULL
)
}
\arguments{
\item{exp}{expression data (if merge==T,exp need a colum to contain gene symbol).}

\item{group}{group can be multi group, and first col is sample name; second col is condition; coef, and multi used to select multi group, keep in mind, control sample priority.}

\item{compared}{compared used to identify which group vs which group (usually <conventional> - <control>). also use for multi group condition.}

\item{normalize}{if the array expression data need to normalized between different sample}

\item{log2}{if the expression data need to be convert to log2}

\item{merge}{if expression data contain same gene symol, you can use it to merge them}

\item{rna.count}{used to execute rna-seq}

\item{p.name}{the colname of pvalue}

\item{fc.name}{the colname of fold change}

\item{p.value}{the threshold of pvalue}

\item{fc.value}{the threshold of fold change}

\item{file.name}{if need to write a txt file to save the deg result, you can assign a file name}
}
\value{
deg result merged with normalized expression data
}
\description{
deg analysis
}
